Finish moving all the classes into the Sierra linker namespace
