+++ sierra_adapter/sierra_items_to_dynamo/src/main/scala/weco/catalogue/sierra_linker/Main.scala
@@ -1,10 +1,9 @@
-package uk.ac.wellcome.platform.sierra_items_to_dynamo
+package weco.catalogue.sierra_linker
 
 import akka.actor.ActorSystem
 import com.typesafe.config.Config
 import org.scanamo.generic.auto._
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient
-import uk.ac.wellcome.platform.sierra_items_to_dynamo.dynamo.Implicits._
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.sns.NotificationMessage
 import uk.ac.wellcome.messaging.typesafe.{SNSBuilder, SQSBuilder}
@@ -13,6 +12,7 @@ import uk.ac.wellcome.storage.store.dynamo.DynamoSingleVersionStore
 import uk.ac.wellcome.storage.typesafe.DynamoBuilder
 import uk.ac.wellcome.typesafe.WellcomeTypesafeApp
 import uk.ac.wellcome.typesafe.config.builders.AkkaBuilder
+import weco.catalogue.sierra_linker.dynamo.Implicits._
 import weco.catalogue.sierra_linker.models.{Link, LinkOps}
 import weco.catalogue.sierra_linker.services.{LinkStore, SierraLinkerWorker}
