+++ sierra_adapter/sierra_items_to_dynamo/src/main/scala/weco/catalogue/sierra_linker/dynamo/Implicits.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.sierra_items_to_dynamo.dynamo
+package weco.catalogue.sierra_linker.dynamo
 
 import org.scanamo.DynamoFormat
 import uk.ac.wellcome.sierra_adapter.model.{SierraBibNumber, SierraItemNumber}
