+++ sierra_adapter/sierra_items_to_dynamo/src/test/scala/weco/catalogue/sierra_linker/SierraLinkerFeatureTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.sierra_items_to_dynamo
+package weco.catalogue.sierra_linker
 
 import org.scalatest.EitherValues
 import org.scalatest.concurrent.{Eventually, IntegrationPatience}
@@ -6,19 +6,13 @@ import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
 import uk.ac.wellcome.messaging.memory.MemoryMessageSender
 import uk.ac.wellcome.sierra_adapter.model.Implicits._
-import uk.ac.wellcome.sierra_adapter.model.{
-  SierraGenerators,
-  SierraHoldingsNumber,
-  SierraHoldingsRecord,
-  SierraItemNumber,
-  SierraItemRecord
-}
+import uk.ac.wellcome.sierra_adapter.model._
 import uk.ac.wellcome.sierra_adapter.utils.SierraAdapterHelpers
 import uk.ac.wellcome.storage.store.memory.MemoryVersionedStore
 import weco.catalogue.sierra_linker.fixtures.WorkerFixture
 import weco.catalogue.sierra_linker.models.Link
 
-class SierraItemsToDynamoFeatureTest
+class SierraLinkerFeatureTest
     extends AnyFunSpec
     with Matchers
     with Eventually
